refactor!: store source byte offsets as u32
Follow-up to #27, which made AST spans u32: every source byte offset in
the crate is now u32 — Token::index(), Error::index(), and the lexer
cursor's internal positions. Token shrinks from 32 to 24 bytes and the
parser's span_index() casting shim is gone.

The 4 GiB source limit moves to Lexer::new so standalone lexer use is
guarded too; Parser::new keeps its assert since it documents the same
contract. The one usize -> u32 conversion point is `source_offset` in
the cursor, used where positions originate from slice lengths and
memchr results.

The cursor's hot byte loops (bump, eatc, name/whitespace runs) switch
from `len`-guarded indexing to `get`-based access: with u32 positions
the optimizer can no longer prove an index in bounds from a u32 length
compare, so guarded indexing would add a bounds check per byte. `get`
keeps it at one check per byte; supergraph_lexer is ~8% faster than
with guarded indexing.

LimitTracker stays usize: it holds token counts and configuration, not
source offsets.

BREAKING CHANGE: Token::index(), Error::index(), and the index
parameters of Error::with_loc/limit/eof are u32; Lexer::new panics on
sources larger than 4 GiB.

+++ crates/oxc_graphql_parser/examples/annotate_snippet.rs
@@ -35,7 +35,7 @@ fn parse_schema() {
         let snippet = Level::ERROR.primary_title(err.message()).element(
             Snippet::source(&src).line_start(0).path(file_name).fold(true).annotation(
                 AnnotationKind::Primary
-                    .span(err.index()..err.index() + err.data().len())
+                    .span(err.index() as usize..err.index() as usize + err.data().len())
                     .label(err.message()),
             ),
         );
